refactor: get rid of legacy EntityStatus
