+++ src/components/EntityStatus/EntityStatus.scss
@@ -1,6 +1,4 @@
-@use '../../styles/mixins';
-
-.ydb-entity-status-new {
+.ydb-entity-status {
     &__note {
         color: inherit;
     }
