+++ src/components/EntityStatus/utils.ts
